+++ dwds/CHANGELOG.md
@@ -4,6 +4,7 @@
   `assetServerPort` and `applicationTarget`.
 - Expose a `BuildRunnerAssetHandler` which proxies request to the asset server
   running within build runner.
+- Account for root directory path when using `package:` URIs with `DartUri`.
 - Support the Legacy Module strategy through the injected client.
 - Support DDK sourcemap URIs.
 - Update SDK dependency to minimum of 2.5.0.